Docker compose (#12)
Using docker compose to setup local databases in order to be able to test things without having to spend too much time on setup.

+++ csharp/Mejram/Models.fs
@@ -1,7 +1,7 @@
 namespace Mejram.Models
 
 type ColumnKey={TableName:string;ColumnName:string}
-type Column={ColumnKey:ColumnKey;Number:int16;ColumnType:string;NotNullConstraint:bool}
+type Column={ColumnKey:ColumnKey;Number:int;ColumnType:string;NotNullConstraint:bool}
 with
   member this.ColumnName=this.ColumnKey.ColumnName
   member this.TableName=this.ColumnKey.TableName
